requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
easyocr>=1.7.0
flask>=2.0.0
pillow>=10.0.0
numpy>=1.24.0 